ai_models/requirements.txt
tensorflow==2.20.0
Pillow==10.3.0
numpy==1.26.4
flask==3.0.0
tensorflow-hub==0.16.1
